Organize supported packages in `mitiq.interface`  (#706)
* Add mitiq.interface, tests passing.

* Update API doc.

* Update docs.

* Ok sphinx.

* Fix style.

* Move conversions to mitiq.interface.

* Fix imports.

* Doh.
