+++ docs/source/apidoc.rst
@@ -1,5 +1,4 @@
-.. mitiq documentation master file, created by
-   sphinx-quickstart on Tue Mar 17 18:16:53 2020.
+.. mitiq documentation master file, created by sphinx-quickstart
 
 API-doc
 =================================
@@ -39,13 +38,33 @@ Clifford Data Regression
 
 See Ref. :cite:`Czarnik_2020_arXiv` for more details on these methods.
 
+Mitiq - Braket
+--------------
+
+Conversions
+^^^^^^^^^^^
+.. automodule:: mitiq.interface.mitiq_braket.conversions
+   :members:
+
+Mitiq - Cirq
+------------
+
+Cirq Utils
+^^^^^^^^^^^
+.. automodule:: mitiq.interface.mitiq_cirq.cirq_utils
+   :members:
 
 Mitiq - PyQuil
 --------------
 
+Conversions
+^^^^^^^^^^^
+.. automodule:: mitiq.interface.mitiq_pyquil.conversions
+   :members:
+
 PyQuil Utils
 ^^^^^^^^^^^^
-.. automodule:: mitiq.mitiq_pyquil.pyquil_utils
+.. automodule:: mitiq.interface.mitiq_pyquil.pyquil_utils
    :members:
 
 Mitiq - Qiskit
@@ -53,12 +72,12 @@ Mitiq - Qiskit
 
 Conversions
 ^^^^^^^^^^^
-.. automodule:: mitiq.mitiq_qiskit.conversions
+.. automodule:: mitiq.interface.mitiq_qiskit.conversions
    :members:
 
 Qiskit Utils
 ^^^^^^^^^^^^
-.. automodule:: mitiq.mitiq_qiskit.qiskit_utils
+.. automodule:: mitiq.interface.mitiq_qiskit.qiskit_utils
    :members:
 
 Probabilistic Error Cancellation
@@ -114,7 +133,7 @@ Noise Scaling: Unitary Folding
 .. automodule:: mitiq.zne.scaling.folding
    :members:
 
-Noise Scaling: Parametric Calibration
+Noise Scaling: Parameter Calibration
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 .. automodule:: mitiq.zne.scaling.parameter
    :members:
