+++ mitiq/interface/mitiq_braket/__init__.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2020 Unitary Fund
+# Copyright (C) 2021 Unitary Fund
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -13,16 +13,4 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-from mitiq.mitiq_qiskit.conversions import (
-    from_qasm,
-    from_qiskit,
-    to_qasm,
-    to_qiskit,
-)
-from mitiq.mitiq_qiskit.qiskit_utils import (
-    execute,
-    execute_with_shots,
-    execute_with_noise,
-    execute_with_shots_and_noise,
-    initialized_depolarizing_noise,
-)
+from mitiq.interface.mitiq_braket.conversions import from_braket, to_braket
