+++ mitiq/interface/mitiq_cirq/__init__.py
@@ -13,4 +13,9 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-from mitiq.mitiq_braket.conversions import from_braket, to_braket
+from mitiq.interface.mitiq_cirq.cirq_utils import (
+    execute,
+    execute_with_shots,
+    execute_with_depolarizing_noise,
+    execute_with_shots_and_depolarizing_noise,
+)
