+++ mitiq/interface/mitiq_pyquil/tests/test_conversions_pyquil.py
@@ -18,7 +18,7 @@ import numpy as np
 from pyquil import Program
 from pyquil.gates import CNOT, CZ, H, RZ, X, Y, Z
 
-from mitiq.mitiq_pyquil.conversions import from_pyquil, to_pyquil
+from mitiq.interface.mitiq_pyquil.conversions import from_pyquil, to_pyquil
 
 
 def test_to_pyquil_from_pyquil_simple():
