+++ mitiq/pec/tests/test_pec.py
@@ -25,7 +25,7 @@ import pyquil
 import qiskit
 
 from mitiq import QPROGRAM
-from mitiq.conversions import convert_to_mitiq, convert_from_mitiq
+from mitiq.interface import convert_to_mitiq, convert_from_mitiq
 from mitiq.benchmarks.utils import noisy_simulation
 
 from mitiq.pec import execute_with_pec, NoisyOperation, OperationRepresentation
